create base class for dash-specific tests
